+++ cmd/maxclaw-gateway/main.go
@@ -1,0 +1,15 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/Lichas/maxclaw/internal/cli"
+)
+
+func main() {
+	if err := cli.ExecuteGateway(); err != nil {
+		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
+		os.Exit(1)
+	}
+}